pyramid: enable auto-instrumentation
