+++ tests/apps/pyramid/pyramid_app/__init__.py
@@ -2,7 +2,7 @@
 # (c) Copyright Instana Inc. 2020
 
 import os
-from tests.apps.pyramid_app.app import pyramid_server as server
+from tests.apps.pyramid.pyramid_app.app import pyramid_server as server
 from tests.apps.utils import launch_background_thread
 
 app_thread = None
